Fix plugins not being updated when updating core
Seems to have regressed in https://github.com/matomo-org/matomo/pull/15770

Haven't tested it yet but this should work. We'll then also need to fix this in Matomo 4 separately because the token_auth check wouldn't work there because it would check against the new app specific token auth table which wouldn't exist yet at the time this is executed.