Added whitespace to the bottom of email template
The issue was that the div element that wrapped the contents of the email
had both top and bottom margins set to zero. There is spacing between
the email text and the (dark blue) header thanks to the 25px bottom margin
on the header. Added a matching 25px bottom margin on the wrapper div.

Fixes #13953.