DB-IP URL is automatically updated in the options
Added a method called by the updater task that updates the DB-IP URL
if needed to the latest downloaded one instead of the one set
in the options when Matomo was initially installed. This removes much of
the confusion at seeing a seemingly outdated URL being used on the GeoIP2
plugin page.

Fixes #16618.