WP optimize is not a matomo plugin (#16820)
Just noticed wordpress found wp-optimize which has a plugin.json but a custom plugin.json and then it was throwing an exception that it's not a valid plugin name because of the dash.

Was only visible on the command line so far so might not be an issue. If this happens again I will find a better generic solution for it.