New version available plugins will be updated confusing (#16890)
We've had 3 issues where people thought that a plugin is not compatible with Matomo 4 because of this screen that is misleading as it shows the current plugin dependencies:
![image](https://user-images.githubusercontent.com/273120/101269833-c6ef4d00-37d7-11eb-8df1-755686435938.png)


For simplicity lets simply only show the name of the plugin as mostly plugins are only compatible with the same major version number anyway and there isn't much value in showing the current required information. Of course we could check if there is actually a version on the marketplace available etc but it would require more work.