Only mark process as finished after fully sending the output (#16952)
Otherwise there could be a race condition in cli multi where it thinks the process is finished, but the output hasn't been sent and written yet. Thus ending up with an "empty response" error under circumstances or only reading a partial output. This should only affect Matomo for WordPress though