Avoid filesystem checks when process was set as finished (#16954)
Because in CliMulti we might check in a loop for every process constantly if they are finished. It won't help all that much but better to not check every time the file system when it is not needed.

There be many other things we could improve there. Will add some comments but probably best not to do them (or we'll see if it's needed later)