Invalidate archives after deleting a visit to prevent race conditions
The deletion of visits can take a while. To prevent any possible race condition we should invalidate archives only after the visits were deleted.

We recently deleted few visits and noticed the reports are still showing the wrong data. Not sure if this was the problem but be still good to change.