Don't try to send security email when no email is provided
Which happens mostly when eg a user is added on CLI.

We noticed this on cloud that it's causing issues and it's trying to add a user, tokens etc. and tries to send emails. The `safeSend` method is catching exceptions however it is logging data as warning. In CLI this causes the return value of an executed CLI command to not be `0` as our console treats any warning as an error.

I checked and generally on CLI the email should be an empty string.

There are also other approaches to fixing this. 

* We could not send the email  on `cli` but generally if there is a user it be better to send it.
* We could develop the check for each individual email but seems like a lot of duplicate code for everything

```
+        $email = Piwik::getCurrentUserEmail();
+        if (!empty($email)) {
+            $mail = $container->make(UserCreatedEmail::class, array(
+                'login' => Piwik::getCurrentUserLogin(),
+                'emailAddress' => Piwik::getCurrentUserEmail(),
+                'userLogin' => $userLogin
+            ));
+            $mail->safeSend();
+        }
```

* We could potentially listen to the "shouldSendMail" event and disable it there in the cloud plugin but this will later again cause issues in core.
* etc.

Generally we wouldn't want to send any such mail while creating a Matomo instance.