Use newer version of Puppeteer for UI tests (#17880)
* Use latest version of Puppeteer for UI tests

* use newer node on travis

* test

* use waitForTimeout / waitForSelector instead of waitFor

* Handle BC for removed waitFor method

* Updates expected UI test files

* update submodules

* improve/fix some UI tests

* Updates expected UI test files

* improve/fix some UI tests

* update url-parse dependency

* simplify some UI tests

* update screenshot

* avoid using merged assets that are broken

* fix expected compare exit code

* don't use merged assets for TwoFactorAuth UI tests

it seems due to too many requests to login, logout and other pages merging the CoreJS assets somehow fails, and results in returning an empty file. Even the reload fails to serve the correct content, so maybe the file, the content should be written to, is somehow locked. Disabling the merged assets seems to fix this issue for this tests. Locally the tests are also running correctly with merged assets.

* improve handling of automatic css / js reloading in UI tests

* increase timeouts of TwoFactorAuth UI tests

* fix 2fa ui test

* revert change

* update new screenshot

* updates UI screenshots

* submodule updates