hack: serialize DeviceDetector and cache it using LazyCache.
The issue is that the constructor DeviceDetector is really really slow, as it is reading and compiling a lot of regular expressions from a yaml file. This is done on almost every tracking request. There is already an in-memory cache, but as each tracking requests starts with a new in-memory cache, this isn't really useful. The DeviceDetector itself has a cache, which prevents evaluating the regular expressions against a user agent string, but it looks like a lot of regular expressions still get compiled.

This shaves off 50% of time per tracking request, getting us from 400req/s to 700req/s or from above 150ms to below 100ms. We're using redis as a cache.

I don't know much php, so this solution is most certainly a dirty hack, but maybe someone can work with this and improve on the solution.