Also unset link token when accepting an invite
Figured this could maybe cause some potential bugs in the future as we almost used something like `$isPendingUser = invite_token is not null or invite_link_token is not null` to check if a user is pending invite. One could assume that if the link token is still set, that the invite is still pending when in reality it isn't and might be better to simply unset it? If not, feel free to close this PR. No problem.