Fix matomo dialog does not close (#20303)
* Fix matomo dialog does not close

Not sure if there's a better solution. It seems this one isn't used anywhere except cloud in the "cancel plan" component.

The component does have a `v-show="modelValue"`, however, because it opens a Modal when it's set to true, this effectively doesn't do anything as the element is somewhere completely different. Calling "close" on the modal will actually close it. Had been debugging in cloud forever until I looked into core and noticed the problem was there.

* built vue files