Updated bug report template: moved Context to the top
The idea of this change is that people can write tickets the way they naturally do (starting with the context) and then answering the point-blank questions like expected behaviour more concisely.
After a few weeks of triaging hundreds of reports, I noticed people often start off their bug reports with a story on what went wrong, where they were etc. This is all context that is currently often put into other sections (such as "Expected behaviour" which was on the top).