align allowed characters for option '--value' on console command config:delete to config:set
The console command `config:set` allows any character for the `--value`
option (see:
`\Piwik\Plugins\CoreAdminHome\Commands\SetConfig\ConfigSettingManipulation::make`).

The regex used for `config:set` is:
`/^([a-zA-Z0-9_]+)\.([a-zA-Z0-9_]+)(\[\])?=(.*)/`

The console command `config:delete` only allows characters a to z (upper
and lower case), numbers and the underscore. This prevents one from
deleting configs values like IPv4 or IPV6 addresses, mail addresses,
networks or IP ranges and lists of values. The currently used regex is:
`/^([a-zA-Z0-9_]+)(?:\.([a-zA-Z0-9_]+))?(?:\[\])?(?:\.([a-zA-Z0-9_]+))?/`

The following is a list of possible settings found in the wild. Those
can be set with `console config:set` but cannot be deleted.

```
[General]
login_allowlist_ip[] = "123.123.123.123"
trusted_hosts = "[2001:858:6::186]:80"
contact_email_address = "no-reply+with-extension@test.at"
proxy_ips[] = "192.168.x.x/24"

[HeatmapSessionRecording]
session_recording_sample_limits = 50,100,250,500,1000,2000,5000
```

This commit aligns the allowed characters for the option `--value` on
the console command `config:delete` to the ones for `config:set` and
changes the regex to
`/^([a-zA-Z0-9_]+)(?:\.([a-zA-Z0-9_]+))?(?:\[\])?(?:\.(.+))?/`.

Fixes: #19966

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>