Don't fetch geolocation database that may not exist yet
When geolocation database updates are configured to be done monthly,
they were scheduled on the 3rd day of _the week_, so Wednesday. For
months that start on a Wednesday, that means that we tried to fetch a
file that might not exist yet.

We now schedule on the 3rd day of _the month_, as I believe was the
original intent of dc98a97182 that introduced the behavior a long time
ago. So this gives 3 days to db-ip.com to update their files in all
cases.

Fixes #18427