Mix and match DDEV+Matomo xhprof setups
Will use the DDEV provided xhprof module, but all profiling controlled
by Matomo using "[Debug] enable_php_profiler=1", CLI parameters, and URL
parameters.

The GUI will be served using the DDEV version, with an alias on the
"/vendor/xhprof" path to keep the required changes to a minimum for now.