Fix invitation links in user invite email template for demo
When inviting a user on demo.matomo.org the link doesn't work because it redirects to demo.matomo.cloud because of the missing index.php. See DEV-19533.

Initially was thinking to add a filter for the slash appendix to be able to reuse it more but decided not worth it for this use case.